backend_chatbot/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sympy==1.12
matplotlib==3.8.2
numpy==1.26.3
networkx==3.2.1
python-multipart==0.0.6
firebase-admin==6.4.0
openai==1.12.0
google-generativeai==0.3.2
pytest==7.4.3
pytest-asyncio==0.23.3
httpx==0.26.0

